requirements.txt
neo4j>=5.14.0
pypdf>=3.17.0
python-docx==1.1.0
openpyxl>=3.1.2
olefile>=0.47
fastapi>=0.104.1
uvicorn>=0.23.2
jinja2>=3.1.2
pytest>=7.4.3
requests>=2.31.0
